+++ stdnet/__init__.py
@@ -2,11 +2,7 @@
 VERSION = (0, 5, 0)
  
 def get_version():
-    if len(VERSION) == 3:
-        v = '%s.%s.%s' % VERSION
-    else:
-        v = '%s.%s' % VERSION[:2]
-    return v
+    return '.'.join(map(str,VERSION))
  
 __version__ = get_version()
 __license__ = "BSD"
@@ -15,7 +11,7 @@ __contact__ = "luca.sbardella@gmail.com"
 __homepage__ = "https://github.com/lsbardel/python-stdnet"
 
 
-sphinxtogithub = False
+sphinxtogithub = True
 
 from .exceptions import *